batch_2/intermediate/flaskProject/requirements.txt
pdfkit==1.0.0
PyPDF2==2.12.1
oss2==2.18.5
urllib3==2.2.1
requests==2.32.3
PyQRCode==1.2.1
mysql-connector-python==8.4.0
pio===0.0.3
oauthlib==3.2.0
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
pypng==0.20220715.0
pillow==10.3.0
python-dotenv==1.0.1
Flask~=3.0.3
qrcode~=7.4.2
config~=0.5.1